requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
cryptography==41.0.5 